removed x86 references, removed Graphics::updateAndRenderFpsTimer as extraneous, configured static analysis rules, moved plan.txt in other $(SolutionDir), fixed Window::saveClipboardTextAsVar, formatting changes, added util::numberToCoordinate, added util::clamp, misc minor lib fixes, moved algorithms.c to utils.c, added magic_enum library for converting enums to string & for other useful enum facilities

+++ KeyEngine/src/blur_outline_draw_pass.cpp
@@ -35,7 +35,7 @@ BlurOutlineDrawPass::BlurOutlineDrawPass( Graphics &gph,
 #if defined _DEBUG && !defined NDEBUG
 	const char *offscreenRtvBlurOutlineName = "OffscreenRenderTargetViewOutlineDraw";
 	m_pRtv->d3dResourceCom()->SetPrivateData( WKPDID_D3DDebugObjectName,
-		strlen( offscreenRtvBlurOutlineName ),
+		(UINT) strlen( offscreenRtvBlurOutlineName ),
 		 offscreenRtvBlurOutlineName );
 #endif
